+++ Content.Server/_KS14/Projectiles/ProjectileLagCompensationComponent.cs
@@ -1,0 +1,4 @@
+namespace Content.Server._KS14.Projectiles;
+
+[RegisterComponent]
+public sealed partial class ProjectileLagCompensationComponent : Component;
