+++ Content.Shared/_Trauma/CCVar/CCvars.Trauma.cs
@@ -6,7 +6,7 @@ namespace Content.Shared._Trauma.CCVar;
 public sealed partial class CCVars
 {
     /// <summary>
-    /// Distance used between projectile and lag-compensated target position for gun prediction.
+    /// Added istance used between projectile and lag-compensated target position for gun prediction.
     /// </summary>
     public static readonly CVarDef<float> GunLagCompRange =
         CVarDef.Create("trauma.gun_lag_comp_range", 0.6f, CVar.SERVER);
